Player updates: Add tests
