+++ src/FplBot.Tests/StateTests.cs
@@ -33,46 +33,7 @@ public class StateTests
         Assert.IsType<FixtureEventsOccured>(_messageSession.PublishedMessages[0].Message);
     }
 
-    [Fact]
-    public async Task WithPriceIncrease()
-    {
-        var state = CreatePriceIncreaseScenario();
-        await state.Reset(1);
-        await state.Refresh(1);
-        Assert.Single(_messageSession.PublishedMessages);
-        Assert.IsType<PlayersPriceChanged>(_messageSession.PublishedMessages[0].Message);
-    }
 
-    [Fact]
-    public async Task WithInjuryUpdate()
-    {
-        var state = CreateNewInjuryScenario();
-        await state.Reset(1);
-        await state.Refresh(1);
-        Assert.Single(_messageSession.PublishedMessages);
-        Assert.IsType<InjuryUpdateOccured>(_messageSession.PublishedMessages[0].Message);
-    }
-
-    [Fact]
-    public async Task WithNewPlayer()
-    {
-        var state = CreateNewPlayerScenario();
-        await state.Reset(1);
-        await state.Refresh(1);
-
-        Assert.Single(_messageSession.PublishedMessages);
-        Assert.IsType<NewPlayersRegistered>(_messageSession.PublishedMessages[0].Message);
-    }
-
-    [Fact]
-    public async Task WithChangeInDoubtfulnessEmitsEvent()
-    {
-        var state = CreateChangeInDoubtfulnessScenario();
-        await state.Reset(1);
-        await state.Refresh(1);
-        Assert.Single(_messageSession.PublishedMessages);
-        Assert.IsType<InjuryUpdateOccured>(_messageSession.PublishedMessages[0].Message);
-    }
 
     [Fact]
     public async Task WithNoFinishedFixtures_DoesNotEmitEvent()
@@ -242,138 +203,7 @@ public class StateTests
         return CreateBaseScenario(fixtureClient, playerClient);
     }
 
-    private static State CreateNewInjuryScenario()
-    {
-        var settingsClient = A.Fake<IGlobalSettingsClient>();
-        A.CallTo(() => settingsClient.GetGlobalSettings()).Returns(new GlobalSettings
-            {
-                Teams = new List<Team>
-                {
-                    TestBuilder.HomeTeam(),
-                    TestBuilder.AwayTeam()
-                },
-                Players = new List<Player>
-                {
-                    TestBuilder.Player().WithStatus(PlayerStatuses.Available)
-                }
-            }
-        ).Once().Then.Returns(new GlobalSettings
-        {
-            Teams = new List<Team>
-            {
-                TestBuilder.HomeTeam(),
-                TestBuilder.AwayTeam()
-            },
-            Players = new List<Player>
-            {
-                TestBuilder.Player().WithStatus(PlayerStatuses.Injured)
-            }
-        });
 
-        var fixtureClient = A.Fake<IFixtureClient>();
-
-        return CreateBaseScenario(fixtureClient, settingsClient);
-    }
-
-    private static State CreateChangeInDoubtfulnessScenario()
-    {
-        var settingsClient = A.Fake<IGlobalSettingsClient>();
-        A.CallTo(() => settingsClient.GetGlobalSettings()).Returns(new GlobalSettings
-            {
-                Teams = new List<Team>
-                {
-                    TestBuilder.HomeTeam(),
-                    TestBuilder.AwayTeam()
-                },
-                Players = new List<Player>
-                {
-                    TestBuilder.Player().WithStatus(PlayerStatuses.Doubtful).WithNews("Knock - 75% chance of playing"),
-                }
-            }
-        ).Once().Then.Returns(new GlobalSettings
-        {
-            Teams = new List<Team>
-            {
-                TestBuilder.HomeTeam(),
-                TestBuilder.AwayTeam()
-            },
-            Players = new List<Player>
-            {
-                TestBuilder.Player().WithStatus(PlayerStatuses.Doubtful).WithNews("Knock - 25% chance of playing")
-            }
-        });
-
-        var fixtureClient = A.Fake<IFixtureClient>();
-
-        return CreateBaseScenario(fixtureClient, settingsClient);
-    }
-
-    private static State CreateNewPlayerScenario()
-    {
-        var settingsClient = A.Fake<IGlobalSettingsClient>();
-        A.CallTo(() => settingsClient.GetGlobalSettings()).Returns(new GlobalSettings
-            {
-                Teams = new List<Team>
-                {
-                    TestBuilder.HomeTeam(),
-                    TestBuilder.AwayTeam()
-                },
-                Players = new List<Player>
-                {
-                    TestBuilder.Player().WithStatus(PlayerStatuses.Available)
-                }
-            }
-        ).Once().Then.Returns(new GlobalSettings
-        {
-            Teams = new List<Team>
-            {
-                TestBuilder.HomeTeam(),
-                TestBuilder.AwayTeam()
-            },
-            Players = new List<Player>
-            {
-                TestBuilder.Player().WithStatus(PlayerStatuses.Available),
-                TestBuilder.OtherPlayer().WithStatus(PlayerStatuses.Available)
-            }
-        });
-
-        var fixtureClient = A.Fake<IFixtureClient>();
-
-        return CreateBaseScenario(fixtureClient, settingsClient);
-    }
-
-    private static State CreatePriceIncreaseScenario()
-    {
-        var playerClient = A.Fake<IGlobalSettingsClient>();
-        A.CallTo(() => playerClient.GetGlobalSettings()).Returns(new GlobalSettings
-            {
-                Teams = new List<Team>
-                {
-                    TestBuilder.HomeTeam(),
-                    TestBuilder.AwayTeam()
-                },
-                Players = new List<Player>
-                {
-                    TestBuilder.Player()
-                }
-            }
-        ).Once().Then.Returns(new GlobalSettings
-        {
-            Teams = new List<Team>
-            {
-                TestBuilder.HomeTeam(),
-                TestBuilder.AwayTeam()
-            },
-            Players = new List<Player>
-            {
-                TestBuilder.Player().WithCostChangeEvent(1)
-            }
-        });
-
-        var fixtureClient = A.Fake<IFixtureClient>();
-
-        return CreateBaseScenario(fixtureClient, playerClient);
-    }
 
     private static State CreateBaseScenario(IFixtureClient fixtureClient, IGlobalSettingsClient settingsClient)
     {
